refactor(dom): split index.ts by subcommand responsibility (#216)
The 733-line dom/index.ts mixed command registration with four distinct
subcommand implementations (query/get/screenshot/eval) plus shared
semantic-node formatting utilities. Navigating or changing any one of
them required paging through the others.

Split into:
- dom/query.ts (36 lines)        — dom query
- dom/get.ts (167 lines)         — dom get (index + selector, raw + semantic)
- dom/screenshot.ts (214 lines)  — page/element/sequence capture
- dom/eval.ts (36 lines)         — dom eval
- dom/semanticUtils.ts (127 lines) — shared a11y+DOM context formatting
- dom/index.ts (90 lines)        — pure command registration

No behavior change. registerDomCommands is now a thin dispatch table.
Each handler file has a single responsibility; changes to one
subcommand no longer touch the others.

Net: -63 lines after removing duplication between query.ts and get.ts
(both use the shared semanticUtils). 620 tests still pass.
